+++ src/dcos_e2e/cluster.py
@@ -15,7 +15,6 @@ from retry import retry
 # Ignore a spurious error - this import is used in a type hint.
 from .backends import ClusterManager  # noqa: F401
 from .backends import ClusterBackend, _ExistingCluster
-from .distributions import Distribution
 from .node import Node
 
 
@@ -47,13 +46,13 @@ class Cluster(ContextDecorator):
                 the installer node before installing DC/OS.
         """
         self._default_ssh_user = cluster_backend.default_ssh_user
+
         self._cluster = cluster_backend.cluster_cls(
             masters=masters,
             agents=agents,
             public_agents=public_agents,
             files_to_copy_to_installer=dict(files_to_copy_to_installer or {}),
             cluster_backend=cluster_backend,
-            linux_distribution=Distribution.CENTOS_7,
         )  # type: ClusterManager
 
     @classmethod