+++ tests/backends/test_docker.py
@@ -13,6 +13,7 @@ from py.path import local  # pylint: disable=no-name-in-module, import-error
 from dcos_e2e.backends import Docker
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.distributions import Distribution
+from dcos_e2e.node import Node
 
 
 class TestDockerBackend:
@@ -111,10 +112,66 @@ class TestDistributions:
     Tests for setting distributions.
     """
 
+    def _get_node_distribution(
+        self,
+        node: Node,
+        default_ssh_user: str,
+    ) -> Distribution:
+        """
+        Given a `Node`, return the `Distribution` on that node.
+        """
+        cat_cmd = node.run(
+            args=['cat /etc/*-release'],
+            user=default_ssh_user,
+            shell=True,
+        )
+
+        version_info = cat_cmd.stdout
+        version_info_lines = [
+            line for line in version_info.decode().split('\n') if '=' in line
+        ]
+        version_data = dict(item.split('=') for item in version_info_lines)
+
+        distributions = {
+            ('"centos"', '"7"'): Distribution.CENTOS_7,
+            ('ubuntu', '"16.04"'): Distribution.UBUNTU_16_04,
+            ('coreos', '12.98.7.0'): Distribution.COREOS,
+            ('fedora', '23'): Distribution.FEDORA_23,
+            ('debian', '"8"'): Distribution.DEBIAN_8,
+        }
+
+        return distributions[(version_data['ID'], version_data['VERSION_ID'])]
+
     def test_default(self) -> None:
         """
-        The default distribution is CentOS 7.
+        The default Linux distribution for a `Node`s is the default Linux
+        distribution of the backend.
         """
-        cluster_backend = Docker()
-        default_distribution = cluster_backend.default_linux_distribution
-        assert default_distribution == Distribution.CENTOS_7
+        with Cluster(
+            cluster_backend=Docker(),
+            masters=1,
+            agents=0,
+            public_agents=0,
+        ) as cluster:
+            (master, ) = cluster.masters
+            node_distribution = self._get_node_distribution(
+                node=master,
+                default_ssh_user=cluster.default_ssh_user,
+            )
+
+        assert node_distribution == Distribution.CENTOS_7
+
+    @pytest.mark.parametrize(
+        'unsupported_linux_distribution',
+        set(Distribution) - {Distribution.CENTOS_7}
+    )
+    def test_custom_choice(
+        self,
+        unsupported_linux_distribution: Distribution,
+    ) -> None:
+        """
+        Starting a cluster with a non-default Linux distribution raises a
+        `NotImplementedError`.
+        """
+        with pytest.raises(NotImplementedError):
+            Docker(linux_distribution=unsupported_linux_distribution)